requirements.txt
langchain-core>=0.3.0
langchain>=0.3.0
langgraph>=0.2.0
langchain-google-genai>=2.0.0
python-dotenv>=1.0.0
google-search-results>=2.4.0
